Time: 0 ms (100%), Space: 42 MB (30.36%) - LeetHub

+++ 0027-remove-element/0027-remove-element.java
@@ -1,7 +1,7 @@
 //4
 class Solution {
     public int removeElement(int[] nums, int val){
-        int j = 0;
+        short j = 0;
         int l = nums.length;
         for(int i=0; i<l; i++){
             if(nums[i] != val){
